+++ src/Core/WorkSpace/src/Components/DataViewerHeader/Header.tsx
@@ -1,11 +1,21 @@
 import { useFactory } from '@equinor/DataFactory';
-import { Tabs } from '@equinor/eds-core-react';
+import { Button, Tabs } from '@equinor/eds-core-react';
 import { useFilteredData } from '@equinor/filter';
+import { DateTime } from 'luxon';
+import { useEffect, useState } from 'react';
+import { ClickableIcon } from '../../../../../components/Icon/ClickableIcon';
 import Icon from '../../../../../components/Icon/Icon';
 import { StatusBar } from '../../../../../packages/StatusBar';
 import { useDataContext } from '../../Context/DataProvider';
 import { TabButton } from '../ToggleButton';
-import { Divider, HeaderWrapper, LeftSection, RightSection, Title } from './HeaderStyles';
+import {
+    Divider,
+    HeaderWrapper,
+    LeftSection,
+    RightSection,
+    Title,
+    TitleHeader,
+} from './HeaderStyles';
 
 const { Tab, List } = Tabs;
 
@@ -29,14 +39,29 @@ export const CompletionViewHeader = ({
     handleFilter,
     activeFilter,
 }: CompletionViewHeaderProps): JSX.Element => {
-    const { statusFunc, key } = useDataContext();
+    const { statusFunc, key, dataApi } = useDataContext();
     const { factory, setSelected } = useFactory(key);
     const { data } = useFilteredData();
+    const [timestamp, setTimestamp] = useState<string | null>(makeTimestamp(dataApi.dataUpdatedAt));
+
+    function makeTimestamp(timeInMs: number): string | null {
+        return DateTime.fromMillis(timeInMs).toRelative({ unit: 'minutes' });
+    }
+
+    useEffect(() => {
+        setInterval(() => setTimestamp(makeTimestamp(dataApi.dataUpdatedAt)), 1000 * 60);
+    }, [dataApi.dataUpdatedAt]);
 
     return (
         <HeaderWrapper>
             <LeftSection>
-                <Title variant="h3">{title}</Title>
+                <TitleHeader>
+                    <Title variant="h3">{title}</Title>
+                    <div>
+                        {`Updated ${timestamp}`}
+                        <ClickableIcon name="refresh" onClick={() => dataApi.refetch()} />
+                    </div>
+                </TitleHeader>
                 {statusFunc && <StatusBar data={statusFunc(data)} />}
             </LeftSection>
             <RightSection>
