+++ src/Core/WorkSpace/src/Context/DataProvider.tsx
@@ -1,9 +1,8 @@
 import { AnalyticsOptions } from '@equinor/Diagrams';
 import { CircularProgress, Icon } from '@equinor/eds-core-react';
 import { FilterOptions } from '@equinor/filter';
-import { useCancellation } from '@equinor/Utils';
 import { createContext, useContext, useEffect, useReducer } from 'react';
-import { useQuery } from 'react-query';
+import { useQuery, useQueryClient, UseQueryResult } from 'react-query';
 import styled from 'styled-components';
 import { ActionType, createCustomAction, getType } from 'typesafe-actions';
 import { GardenOptions } from '../../../../components/ParkView/Models/gardenOptions';
@@ -14,14 +13,13 @@ import {
     TableOptions,
     TreeOptions,
     WorkflowEditorOptions,
-} from '../WorkSpaceApi/State';
+} from '../WorkSpaceApi/workspaceState';
 import { useWorkSpace } from '../WorkSpaceApi/useWorkSpace';
 import { DataViewerProps, ViewOptions } from '../WorkSpaceApi/WorkSpaceTypes';
 
 interface DataState {
     key: string;
     name: string;
-    data: any[];
     subData: Record<string, any[]>;
     item: Record<string, unknown>;
     viewComponent?: React.FC<DataViewerProps<unknown>>;
@@ -37,23 +35,28 @@ interface DataState {
     workflowEditorOptions?: WorkflowEditorOptions;
 }
 interface DataContextState extends DataState {
-    getData: VoidFunction;
-    isLoading: boolean;
-    error: unknown;
+    data: any[];
+    dataApi: DataApi;
 }
+
+type DataApi = DataOperations & UseQueryResult<unknown[] | undefined, unknown>;
+
+interface DataOperations {
+    patchRecord: (id: string, item: unknown, identifier?: string) => void;
+    deleteRecord: (id: string, identifier?: string) => void;
+    getRecord: (id: string, identifier?: string) => void;
+    insertRecord: (item: unknown) => void;
+}
+
 interface DataProviderProps {
     children: React.ReactNode;
 }
 
-type VoidFunction = () => void;
-
 export enum DataAction {
-    getData = 'getData',
     setOptions = 'setOptions',
 }
 
 export const actions = {
-    getData: createCustomAction(DataAction.getData, (data: any[]) => ({ data })),
     setOptions: createCustomAction(DataAction.setOptions, (options) => ({ options })),
 };
 
@@ -65,10 +68,8 @@ const DataContext = createContext({} as DataContextState);
 
 export function ClientReducer(state: DataState, action: Action): DataState {
     switch (action.type) {
-        case getType(actions.getData):
-            return { ...state, data: action.data };
         case getType(actions.setOptions):
-            return { ...state, ...action.options, data: [] };
+            return { ...state, ...action.options };
         default:
             return state;
     }
@@ -82,57 +83,91 @@ export const DataProvider = ({ children }: DataProviderProps): JSX.Element => {
 
     const initialState: DataState = {
         key,
-        data: [],
         subData: {},
         item: {},
         ...options,
     };
-    const { abortController, viewId } = useCancellation(key);
+
     const [state, dispatch] = useReducer(ClientReducer, initialState);
 
     useEffect(() => {
         dispatch(actions.setOptions(initialState));
+        dataApi.remove();
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [key]);
 
-    const { refetch, data, isLoading, error } = useQuery(
+    const dataApi = useQuery(
         key,
         async () => {
             if (!dataSource) return;
-
-            const data = await dataSource(abortController);
-            if (viewId !== key) {
-                return;
-            }
-            dispatch(actions.getData(data));
-            return data;
+            return await dataSource();
         },
-        { refetchOnWindowFocus: false }
+        { refetchOnWindowFocus: false, staleTime: 1 * 1000 * 60 * 60 }
     );
 
     useEffect(() => {
-        if (!data) return;
-        dispatch(actions.getData(data));
-    }, [data]);
-
-    useEffect(() => {
-        refetch();
+        dataApi.refetch();
     }, [dataSource]);
+
+    const queryClient = useQueryClient();
+
+    function patchRecord(id: string, item: unknown, identifier?: string) {
+        const query = queryClient.getQueryCache().find(key);
+        if (!query || !dataApi.data) return;
+
+        const patchIndex = dataApi.data.findIndex(
+            (record: any) => record[identifier ?? options.objectIdentifier] === id
+        );
+        if (patchIndex === -1) return;
+
+        query.setData((dataApi.data[patchIndex] = item));
+    }
+
+    function deleteRecord(id: string, identifier?: string) {
+        const query = queryClient.getQueryCache().find(key);
+        if (!query || !dataApi.data) return;
+
+        query.setData(
+            dataApi.data.filter(
+                (record: any) => record[identifier ?? options.objectIdentifier] !== id
+            )
+        );
+    }
+
+    function getRecord(id: string, identifier?: string) {
+        if (!dataApi.data) return;
+
+        return dataApi.data.find(
+            (record: any) => record[identifier ?? options.objectIdentifier] === id
+        );
+    }
+
+    function insertRecord(item: unknown) {
+        const query = queryClient.getQueryCache().find(key);
+        if (!query || !dataApi.data) return;
+
+        query.setData([item, ...dataApi.data]);
+    }
 
     return (
         <DataContext.Provider
             value={{
                 ...state,
-                getData: refetch,
-                isLoading,
-                error,
+                data: dataApi.data || [],
+                dataApi: {
+                    ...dataApi,
+                    insertRecord,
+                    getRecord,
+                    deleteRecord,
+                    patchRecord,
+                },
             }}
         >
-            {isLoading ? (
+            {dataApi.isLoading ? (
                 <Loading>
                     <CircularProgress value={0} size={48} />
                 </Loading>
-            ) : error ? (
+            ) : dataApi.error ? (
                 <Loading>
                     <Icon name="error_outlined" />
                     Something went wrong
