+++ src/apps/ScopeChangeRequest/WorkspaceConfig/dataCreatorConfig.ts
@@ -1,0 +1,7 @@
+import { FactoryOptions } from '../../../Core/WorkSpace/src';
+import { ScopeChangeRequestForm } from '../Components/Form/ScopeChangeRequestForm';
+
+export const dataCreator: FactoryOptions = {
+    title: 'Scope change',
+    component: ScopeChangeRequestForm,
+};
