+++ src/apps/apps.ts
@@ -155,11 +155,10 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.ProjectInformation,
         icon: '',
-        uri: `${
-            isProduction()
+        uri: `${isProduction()
                 ? 'https://fusion.equinor.com/apps/pro-org/3cf72ff9-c50f-4e94-ba79-31721ba42dec/chart'
                 : 'https://pro-s-portal-ci.azurewebsites.net/apps/pro-org/3cf72ff9-c50f-4e94-ba79-31721ba42dec/chart'
-        }`,
+            }`,
         appEnv: 'prod',
         tags: [],
     },
@@ -192,11 +191,10 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.SSU,
         icon: '',
-        uri: `${
-            isProduction()
+        uri: `${isProduction()
                 ? 'https://fusion.equinor.com/apps/bmt/65728fee-185d-4a0c-a91d-8e3f3781dad8'
                 : 'https://pro-s-portal-ci.azurewebsites.net/apps/bmt/b6552a8f-9173-416f-9fc0-996387ff7e3a'
-        }`,
+            }`,
         appEnv: 'prod',
         tags: ['Fusion', 'Link'],
     },
@@ -304,7 +302,7 @@ export const apps: AppManifest[] = [
         groupe: Apps.ConstructionAndCommissioning,
         icon: '',
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
             setup: WorkOrderSetup,
         },
         tags: ['Job'],
@@ -325,7 +323,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
         },
     },
     {
@@ -336,7 +334,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
             setup: handoverSetup,
         },
         appEnv: 'dev',
@@ -349,7 +347,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
             setup: heatTraceInstallationSetup,
         },
         appEnv: 'dev',
@@ -362,7 +360,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
         },
     },
     {
@@ -373,7 +371,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
         },
     },
     {
@@ -382,9 +380,8 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.ConstructionAndCommissioning,
         icon: '',
-        uri: `https://${
-            isProduction() ? 'procosys' : 'procosystest'
-        }.equinor.com/JOHAN_CASTBERG/Preservation`,
+        uri: `https://${isProduction() ? 'procosys' : 'procosystest'
+            }.equinor.com/JOHAN_CASTBERG/Preservation`,
         tags: ['link', 'procosys'],
         appEnv: 'prod',
     },
@@ -396,7 +393,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
         },
     },
     {
@@ -407,7 +404,7 @@ export const apps: AppManifest[] = [
         icon: '',
         tags: [],
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
             setup: SwcrSetup,
         },
         appEnv: 'dev',
@@ -418,9 +415,8 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.ConstructionAndCommissioning,
         icon: '',
-        uri: `https://${
-            isProduction() ? 'fusion.equinor.com' : 'pro-s-portal-ci.azurewebsites.net'
-        }/apps/dcp`,
+        uri: `https://${isProduction() ? 'fusion.equinor.com' : 'pro-s-portal-ci.azurewebsites.net'
+            }/apps/dcp`,
         tags: ['link', 'fusion'],
         appEnv: 'prod',
     },
@@ -430,9 +426,8 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.ConstructionAndCommissioning,
         icon: '',
-        uri: `https://${
-            isProduction() ? 'procosys' : 'procosystest'
-        }.equinor.com/JOHAN_CASTBERG/InvitationForPunchOut`,
+        uri: `https://${isProduction() ? 'procosys' : 'procosystest'
+            }.equinor.com/JOHAN_CASTBERG/InvitationForPunchOut`,
         tags: ['link', 'procosys'],
         appEnv: 'prod',
     },
@@ -461,7 +456,7 @@ export const apps: AppManifest[] = [
         groupe: Apps.ProjectControl,
         icon: '',
         app: {
-            appType: 'DataViewer',
+            appType: 'Workspace',
             setup: scopeChangeSetup,
         },
         tags: [],
@@ -473,11 +468,10 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.ProjectControl,
         icon: '',
-        uri: `${
-            isProduction()
+        uri: `${isProduction()
                 ? 'https://fusion.equinor.com/apps/management-of-change/3380fe7d-e5b7-441f-8ce9-a8c3133ee499'
                 : 'https://pro-s-portal-ci.azurewebsites.net/apps/management-of-change'
-        }`,
+            }`,
         tags: ['Link', 'Fusion'],
         appEnv: 'prod',
     },
@@ -487,11 +481,10 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.ProjectControl,
         icon: '',
-        uri: `${
-            isProduction()
+        uri: `${isProduction()
                 ? 'https://fusion.equinor.com/apps/project-control-and-analysis/3380fe7d-e5b7-441f-8ce9-a8c3133ee499'
                 : 'https://pro-s-portal-ci.azurewebsites.net/apps/project-control-and-analysis/b9a3246a-ddb5-4086-b4ec-dd4b0e88b700'
-        }`,
+            }`,
         tags: ['Link', 'Fusion'],
         appEnv: 'prod',
     },
@@ -556,11 +549,10 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.Collaboration,
         icon: 'tag',
-        uri: `${
-            isProduction()
+        uri: `${isProduction()
                 ? 'https://fusion.equinor.com/apps/meetings'
                 : 'https://pro-s-portal-ci.azurewebsites.net/apps/meetings'
-        } `,
+            } `,
         tags: ['fuison', 'link', 'external'],
         appEnv: 'prod',
     },
@@ -570,11 +562,10 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.Collaboration,
         icon: 'tag',
-        uri: `${
-            isProduction()
+        uri: `${isProduction()
                 ? 'https://fusion.equinor.com/apps/reviews/255d8c0a-7893-4c21-ab42-62c652ea8129'
                 : 'https://pro-s-portal-ci.azurewebsites.net/apps/reviews'
-        }`,
+            }`,
         tags: ['fuison', 'link', 'external'],
         appEnv: 'prod',
     },
@@ -597,9 +588,8 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.Collaboration,
         icon: '',
-        uri: `https://${
-            isProduction() ? 'procosys' : 'procosystest'
-        }.equinor.com/JOHAN_CASTBERG/Search?searchType=Query`,
+        uri: `https://${isProduction() ? 'procosys' : 'procosystest'
+            }.equinor.com/JOHAN_CASTBERG/Search?searchType=Query`,
         tags: ['link', 'procosys'],
         appEnv: 'prod',
     },
@@ -623,9 +613,8 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.AssetData,
         icon: '',
-        uri: `https://${
-            isProduction() ? 'stid' : 'stidtest'
-        }.equinor.com/JCA/search?type=doc&revstatus=OF%2CUA%2CRE%2CPL%2COF-P`,
+        uri: `https://${isProduction() ? 'stid' : 'stidtest'
+            }.equinor.com/JCA/search?type=doc&revstatus=OF%2CUA%2CRE%2CPL%2COF-P`,
         tags: ['3D', 'Asset', 'Map', 'Doc'],
         appEnv: 'prod',
     },
@@ -635,9 +624,8 @@ export const apps: AppManifest[] = [
         color: '#0364B8',
         groupe: Apps.AssetData,
         icon: 'tag',
-        uri: `https://${
-            isProduction() ? 'stid' : 'stidtest'
-        }.equinor.com/JCA/search?type=tag&tagstatus=A%2CP%2CR%2CF`,
+        uri: `https://${isProduction() ? 'stid' : 'stidtest'
+            }.equinor.com/JCA/search?type=tag&tagstatus=A%2CP%2CR%2CF`,
         tags: ['Tag', 'Data', 'Functional Location'],
         appEnv: 'prod',
     },
