+++ src/components/Layouts/MainLayout.tsx
@@ -4,6 +4,8 @@ import { getWidth } from '../../packages/Sidesheet/Utils/getWidth';
 import { FullscreenMainMenu } from '../Menu/FullscreenMainMenu';
 import { MainMenu } from '../Menu/MainMenu';
 import { ChildrenWrapper, MainMenuWrapper, Wrapper } from './MainLayoutStyles';
+import { ReactQueryDevtools } from 'react-query/devtools';
+import { useGlobalClientState } from '../../Core/Client/ClientState/ClientState';
 
 interface MainLayoutProps {
     children: React.ReactNode;
@@ -14,6 +16,9 @@ export const MainLayout = ({ children }: MainLayoutProps): JSX.Element => {
         settings: { fullscreenMenuActive },
     } = useClientContext();
     const sideSheet = useSideSheet();
+    const {
+        settings: { clientEnv },
+    } = useGlobalClientState();
 
     return (
         <Wrapper>
@@ -22,6 +27,7 @@ export const MainLayout = ({ children }: MainLayoutProps): JSX.Element => {
                 <MainMenu />
             </MainMenuWrapper>
             <ChildrenWrapper sideSheetWidth={getWidth(sideSheet)}>{children}</ChildrenWrapper>
+            {clientEnv === 'dev' && <ReactQueryDevtools initialIsOpen={false} />}
         </Wrapper>
     );
 };
