+++ src/components/TopBar/DevBar/DevBar.tsx
@@ -7,10 +7,12 @@ interface DevBarProps {
 export const DevBar = ({ env }: DevBarProps): JSX.Element | null => {
     if (env === 'test' || env === 'dev')
         return (
-            <BetaTag>
-                <b>UNDER DEVELOPMENT - {env.toUpperCase()} </b>
-                <p>This site contains test data.</p>
-            </BetaTag>
+            <>
+                <BetaTag>
+                    <b>UNDER DEVELOPMENT - {env.toUpperCase()} </b>
+                    <p>This site contains test data.</p>
+                </BetaTag>
+            </>
         );
 
     return null;
